requirements.txt
# Core dependencies
streamlit==1.32.0
openai-whisper==20231117
anthropic==0.18.1
deepgram-sdk==3.1.0
python-dotenv==1.0.1

# Audio processing
pydub==0.25.1
ffmpeg-python==0.2.0

# Cross-platform audio recording
pyaudio==0.2.14  # For Windows
sounddevice==0.4.6  # Alternative for cross-platform
numpy==1.26.4  # Required by sounddevice

# Utilities
tqdm==4.66.2
requests==2.31.0

# Testing
pytest==8.0.0
pytest-env==1.1.3
pytest-cov==4.1.0

# Linting and formatting
flake8==7.0.0
black==24.2.0
isort==5.13.2
